Cleanup and changes based off of TA comments Split styles into their own files, index only imports app as entry point, router and routes moved to app, created homepage
