+++ src/index.js
@@ -1,27 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styles/index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import "./style.css"
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Engineering from './departments/Engineering';
-import Login from './components/Login';
-import AdminPage from './departments/AdminPage';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Router>
-      <Routes>
-        <Route path='/' element={<App/>} />
-        <Route exact path='/Engineering' element={<Engineering/>} />
-        <Route exact path='/Login' element={<Login/>} />
-        <Route exact path='/AdminPage' element={<AdminPage/>} />
-      </Routes>
-    </Router>
+    <App />
   </React.StrictMode>
 );
+
+//  **TODO**
+// index should only relate to whtever is created in the DOM, this should only himport APP as the entrypoint, ROUTER should go into APP
+//
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
