--- src/style.css
@@ -1,150 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family:'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
-    font-size: large;
-    display: flex;
-    height: 100vh;
-    width: 100%;
-    padding: 0 10px;
-}
-
-#root {
-    flex-grow: 1;
-    display: flex;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.sidebar{
-    border-right: 1px solid black;
-    flex: 0.2;
-    padding-left: 20px;
-    padding-right: 20px;
-    height: 100vh;
-}
-
-.sidebar img {
-    height: 90px;
-    width: auto;
-}
-
-.sidebar a{
-    text-decoration: none;
-    color: inherit;
-}
-
-.departments{
-    display: flex;
-    align-items: center;
-    padding: 20px;
-    cursor: pointer;
-}
-
-.departments h2 {
-    font-weight: 800;
-    font-size: 20px;
-    margin-right: 20px;
-}
-
-
-.departments:hover{
-    background-color: #FFCCCC;
-    border-radius: 30px;
-    transition: color 100ms ease-out;
-}
-
-.departments.active {
-    background-color: #FFCCCC;
-    border-radius: 30px;
-}
-
-.feed{
-    flex: 1;
-    position: sticky;
-    top: 0;
-}
-
-
-.title {
-    position: sticky;
-    top: 0;
-    padding: 20px;
-    border-bottom: 1px solid black;
-    background-color: white;
-}
-.title h2{
-    display: flex;
-    justify-content: flex-start;
-}
-
-.title button{
-    background: #FFF;
-    border: 1px solid;
-    padding: .5rem 1.1rem;
-    cursor: pointer;
-    border-radius: .25rem;
-    float: right;
-}
-
-.title button:hover {
-    background: #FFCCCC;
-    transition: all 0.35s ease;
-}
-
-.post_profile_pic img{
-    height: 60px;
-    padding: 10px;
-}
-
-.post{
-    display: flex;
-    align-items: left;
-    border-bottom: 1px solid black;
-    padding: 10px;
-}
-
-.post_content img{
-    width: 200px;
-    object-fit: contain;
-    border-radius: 20px;
-}
-
-.post_content {
-    padding: 10px;
-}
-
-.post_text_container h3{
-    font-size: 20px;
-    margin-bottom: 5px;
-}
-
-.post_description{
-    margin-bottom: 10px;
-    font-size: 20px;
-}
-
-.App{
-    display: flex;
-    height: 100vh;
-    flex-grow: 1;
-    text-align: left;
-}
-
-.Left-side{
-    float: left;
-}
-
-.News-Feed{
-    float: right;
-    flex-grow: 1;
-    overflow-y: scroll;
-}
-
-.News-Feed::-webkit-scrollbar{
-    display: none;
-}
