+++ src/styles/App.css
@@ -1,5 +1,8 @@
-.App {
-  text-align: center;
+.App{
+  display: flex;
+  height: 100vh;
+  flex-grow: 1;
+  text-align: left;
 }
 
 .App-logo {
