+++ src/styles/Newsfeed.css
@@ -1,0 +1,31 @@
+
+.title {
+    position: sticky;
+    top: 0;
+    padding: 20px;
+    border-bottom: 1px solid black;
+    background-color: white;
+}
+
+.title h2{
+    display: flex;
+    justify-content: flex-start;
+}
+
+.title h2:hover{
+    cursor: pointer;
+}
+
+.title button{
+    background: #FFF;
+    border: 1px solid;
+    padding: .5rem 1.1rem;
+    cursor: pointer;
+    border-radius: .25rem;
+    float: right;
+}
+
+.title button:hover {
+    background: #FFCCCC;
+    transition: all 0.35s ease;
+}
